requirements.txt
dify-plugin==0.0.1b72
azure-ai-documentintelligence>=1.0.0
python-dotenv>=0.19.0
azure-core>=1.26.0
requests>=2.31.0
websocket-client>=1.6.4
python-dotenv>=1.0.0
Pillow>=10.2.0
PyMuPDF>=1.23.26
PyPDF2>=3.0.1
